requirements.txt
crewai
openai
duckduckgo-search
spacy
sentence-transformers
chromadb
fastapi
uvicorn[standard]
python-dotenv
requests
pandas
numpy
pytest
pytest-asyncio
python-multipart